Move the active resolve attempts class to a separate header/cpp file for better code organization
