+++ src/lib/mdns/Resolver_ImplMinimalMdns.cpp
@@ -24,6 +24,7 @@
 #include <lib/mdns/MinimalMdnsServer.h>
 #include <lib/mdns/ServiceNaming.h>
 #include <lib/mdns/TxtFields.h>
+#include <lib/mdns/minimal/ActiveResolveAttempts.h>
 #include <lib/mdns/minimal/Parser.h>
 #include <lib/mdns/minimal/QueryBuilder.h>
 #include <lib/mdns/minimal/RecordData.h>
@@ -65,192 +66,6 @@ constexpr uint16_t kMdnsPort        = 5353;
 
 using namespace mdns::Minimal;
 using MdnsCacheType = Mdns::MdnsCache<CHIP_CONFIG_MDNS_CACHE_SIZE>;
-
-/// Keeps track of active resolve attempts
-class ActiveResolveAttempts
-{
-public:
-    ActiveResolveAttempts() { Reset(); }
-
-    void Reset()
-    {
-        for (auto & item : mRetryQueue)
-        {
-            item.peerId.SetNodeId(kUndefinedNodeId);
-        }
-    }
-
-    void Complete(const PeerId & peerId)
-    {
-        for (auto & item : mRetryQueue)
-        {
-            if (item.peerId == peerId)
-            {
-                item.peerId.SetNodeId(kUndefinedNodeId);
-                return;
-            }
-        }
-
-        // This may happen during boot time adverisements: nodes come online
-        // and advertise their IP without any explicit queries for them
-        ChipLogProgress(Discovery, "Discovered node without a pending query");
-    }
-
-    void MarkPending(const PeerId & peerId)
-    {
-        // Strategy when picking the peer id to use:
-        //   1 if a matching peer id is already found, use that one
-        //   2 if an 'unused' entry is found, use that
-        //   3 otherwise expire the one with the largest nextRetryDelaySec
-        //     or if equal nextRetryDelaySec, pick the one with the oldest
-        //     queryDueTimeMs
-
-        RetryEntry * entryToUse = &mRetryQueue[0];
-
-        for (size_t i = 1; i < kRetryQueueSize; i++)
-        {
-            if (entryToUse->peerId == peerId)
-            {
-                break; // best match possible
-            }
-
-            RetryEntry * entry = mRetryQueue + i;
-
-            // Rule 1: peer id match always matches
-            if (entry->peerId == peerId)
-            {
-                entryToUse = entry;
-                continue;
-            }
-
-            // Rule 2: select unused entries
-            if ((entryToUse->peerId.GetNodeId() != kUndefinedNodeId) && (entry->peerId.GetNodeId() == kUndefinedNodeId))
-            {
-                entryToUse = entry;
-                continue;
-            }
-            else if (entryToUse->peerId.GetNodeId() == kUndefinedNodeId)
-            {
-                continue;
-            }
-
-            // Rule 2: both choices are used (have a defined node id),
-            // pick the one with most queryDueTimeMs
-            if (entry->nextRetryDelaySec > entryToUse->nextRetryDelaySec)
-            {
-                entryToUse = entry;
-            }
-            else if ((entry->nextRetryDelaySec == entryToUse->nextRetryDelaySec) &&
-                     (entry->queryDueTimeMs < entryToUse->queryDueTimeMs))
-            {
-                entryToUse = entry;
-            }
-        }
-
-        if ((entryToUse->peerId.GetNodeId() != kUndefinedNodeId) && (entryToUse->peerId != peerId))
-        {
-            ChipLogError(Discovery, "Re-using pending resolve entry before reply was received.");
-        }
-
-        entryToUse->peerId            = peerId;
-        entryToUse->queryDueTimeMs    = chip::System::Clock::GetMonotonicMilliseconds();
-        entryToUse->nextRetryDelaySec = 1;
-    }
-
-    static constexpr int kInvalidNextMs = 1;
-
-    // Get minimum milliseconds until the next pending reply is required.
-    //
-    // Returns -1 if no active resolve entries exist, >= 0 if a delay is expected
-    int GetMsUntilNextExpectedResponse() const
-    {
-        int minDelay = kInvalidNextMs;
-
-        chip::System::Clock::MonotonicMilliseconds nowMs = System::Clock::GetMonotonicMilliseconds();
-
-        for (auto & entry : mRetryQueue)
-        {
-            if (entry.peerId.GetNodeId() == kUndefinedNodeId)
-            {
-                continue;
-            }
-
-            if (nowMs >= entry.queryDueTimeMs)
-            {
-                minDelay = 0;
-                break;
-            }
-
-            minDelay = std::min(minDelay, static_cast<int>(entry.queryDueTimeMs - minDelay));
-        }
-
-        return minDelay;
-    }
-
-    // Get the peer Id that needs scheduling for a query
-    //
-    // Assumes that the resolution is being sent and will apply internal
-    // query logic.
-    //
-    // Returns kUndefinedNodeId if no peer scheduled.
-    PeerId NextScheduledPeer()
-    {
-        chip::System::Clock::MonotonicMilliseconds nowMs = System::Clock::GetMonotonicMilliseconds();
-
-        for (auto & entry : mRetryQueue)
-        {
-            if (entry.peerId.GetNodeId() == kUndefinedNodeId)
-            {
-                continue; // not a pending item
-            }
-
-            if (entry.queryDueTimeMs > nowMs)
-            {
-                continue; // not yet due
-            }
-
-            if (entry.nextRetryDelaySec > kMaxRetryDelaySec)
-            {
-                ChipLogError(Discovery, "Timeout waiting for mDNS resolution.");
-                entry.peerId.SetNodeId(kUndefinedNodeId);
-                continue;
-            }
-
-            entry.queryDueTimeMs = nowMs + entry.nextRetryDelaySec * 1000L;
-            entry.nextRetryDelaySec *= 2;
-            return entry.peerId;
-        }
-        return PeerId();
-    }
-
-private:
-    static constexpr uint32_t kMaxRetryDelaySec = 16;
-
-    struct RetryEntry
-    {
-        // What peer id is pending discovery.
-        //
-        // Inactive entries are marked by having NodeId == kUndefinedNodeId
-        PeerId peerId;
-
-        // When a reply is expected for this item
-        chip::System::Clock::MonotonicMilliseconds queryDueTimeMs;
-
-        // Next expected delay for sending if reply is not reached by
-        // 'queryDueTimeMs'
-        //
-        // Based on RFC 6762 expectations are:
-        //    - the interval between the first two queries MUST be at least
-        //      one second
-        //    - the intervals between successive queries MUST increase by at
-        //      least a factor of two
-        uint32_t nextRetryDelaySec = 1;
-    };
-
-    static constexpr size_t kRetryQueueSize = 4;
-
-    RetryEntry mRetryQueue[kRetryQueueSize];
-};
 
 class PacketDataReporter : public ParserDelegate
 {
