+++ src/lib/mdns/minimal/BUILD.gn
@@ -16,6 +16,8 @@ import("//build_overrides/chip.gni")
 
 static_library("minimal") {
   sources = [
+    "ActiveResolveAttempts.h",
+    "ActiveResolveAttempts.cpp",
     "Parser.cpp",
     "Parser.h",
     "Query.h",
